add tests for factories
