+++ lib/pgdice/partition_manager_factory.rb
@@ -7,11 +7,11 @@ module PgDice
 
     def initialize(configuration, opts = {})
       @configuration = configuration
-      @logger_factory = opts[:logger_factory] ||= proc {@configuration.logger}
-      @batch_size_factory = opts[:batch_size_factory] ||= proc {@configuration.batch_size}
-      @approved_tables_factory = opts[:approved_tables_factory] ||= proc {@configuration.approved_tables}
+      @logger_factory = opts[:logger_factory] ||= proc { @configuration.logger }
+      @batch_size_factory = opts[:batch_size_factory] ||= proc { @configuration.batch_size }
+      @approved_tables_factory = opts[:approved_tables_factory] ||= proc { @configuration.approved_tables }
       @validation_factory = opts[:validation_factory] ||= PgDice::ValidationFactory.new(configuration)
-      @partition_adder_factory = opts[:partition_adder_factory] ||= PgDice::PgSliceManagerFactory.new(configuration)
+      @partition_adder_factory = opts[:partition_adder_factory] ||= partition_adder_factory
       @partition_lister_factory = opts[:partition_lister_factory] ||= PgDice::PartitionListerFactory.new(configuration)
       @partition_dropper_factory = opts[:partition_dropper_factory] ||= PgDice::PartitionDropperFactory.new(configuration)
     end
@@ -24,6 +24,15 @@ module PgDice
                                    partition_adder: @partition_adder_factory.call,
                                    partition_lister: @partition_lister_factory.call,
                                    partition_dropper: @partition_dropper_factory.call)
+    end
+
+    private
+
+    def partition_adder_factory
+      proc do
+        pg_slice_manager = PgDice::PgSliceManagerFactory.new(@configuration).call
+        lambda { |all_params| pg_slice_manager.add_partitions(all_params) }
+      end
     end
   end
 end
