+++ test/pgdice/functionality_test.rb
@@ -114,6 +114,33 @@ class FunctionalityTest < Minitest::Test
                                             .list_droppable_partitions(table_name, older_than: tomorrow).size
   end
 
+
+  def test_works_year_tables
+    table_name = 'posts'
+    PgDice.partition_helper.partition_table!(table_name, past: 1, future: 0, period: :year)
+    PgDice.partition_manager.add_new_partitions(table_name, past: 2, future: 2, period: :year)
+
+    PgDice.validation.assert_tables(table_name, future: 2, past: 2)
+
+    assert_future_tables_error { PgDice.validation.assert_tables(table_name, future: 3) }
+    assert_past_tables_error { PgDice.validation.assert_tables(table_name, past: 3) }
+  ensure
+    partition_helper.undo_partitioning('posts')
+  end
+
+  def test_works_month_tables
+    table_name = 'posts'
+    PgDice.partition_helper.partition_table!(table_name, past: 1, future: 0, period: :month)
+    PgDice.partition_manager.add_new_partitions(table_name, future: 2, past: 2, period: :month)
+
+    PgDice.validation.assert_tables(table_name, future: 2, past: 2)
+
+    assert_future_tables_error { PgDice.validation.assert_tables(table_name, future: 3) }
+    assert_past_tables_error { PgDice.validation.assert_tables(table_name, past: 3) }
+  ensure
+    partition_helper.undo_partitioning('posts')
+  end
+
   private
 
   def batch_size_and_minimum_tables
