+++ test/pgdice/partition_lister_factory_test.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class PartitionListerFactoryTest < Minitest::Test
+  def test_create_partition_lister
+    assert PgDice::PartitionListerFactory.new(PgDice.configuration).call
+  end
+end
