+++ test/pgdice/partition_lister_test.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class PartitionListerTest < Minitest::Test
+
+end
