+++ test/pgdice/pg_slice_manager_factory_test.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class PgSliceManagerFactoryTest < Minitest::Test
+  def test_create_pg_slice_manager
+    assert PgDice::PgSliceManagerFactory.new(PgDice.configuration).call
+  end
+end
